Slider feature

+++ database/migrations/2019_09_03_121749_create_sections_table.php
@@ -16,7 +16,7 @@ class CreateSectionsTable extends Migration
         Schema::create('sections', function (Blueprint $table) {
             $table->bigIncrements('id');
 
-            $table->integer('course_id')->unsigned()->nullable();
+            $table->bigInteger('course_id')->unsigned()->nullable();
             $table->foreign('course_id')->references('id')->on('courses');
             $table->string('title')->nullable();
             $table->integer('position')->nullable()->unsigned();
